CHANGE FUNCTION FOR CHECKING WINNER OF GAME
I have refactored the function for checking for a winner. First changed the game IIFE to include a getCell and setCell function. Further, I restructured the checkWinner function that will check for the winning combos and log the winner after three rounds. I also changed the playGame function. The game functions well in the console now.