ADD RESET GAME FUNCTION TO GAMEFLOW FUNCTION
I have included a gameOver variable to gameFlow function and given it a vaue of false. Further, I have added a function to check whether the gameboard is full and restructured my code in playGame function to check whether the game is over and the board is full. I also created a resetGame function that resets the game when the game is over.