ADD DISPLAY CONTROLLER FUNCTION
I have added a function that controls the display of the game board on the webpage. Further, I have added an event listener to listen for clicks on individual cells of the gameboard by each player. I have further styled the board in order to see it on the webpage.