ADD DIALOG FORM TO HTML
I have added a dialog form to the html file to take player name and marker inputs. I have further restractured the javascript file in order to properly store the player names and markers in an object and further display the player names and markers on the webpage. The game is functioning and player names and markers are displayed on screen.