REFACTOR UPDATESCREEN FUNCTION.
I have refactored updateScreen() function by looping through the cells and updating only their content and further creating a new function renderInitialBoard that is returned when init() is run. This new function creates the board itself. This refactoring is required in order to properly render the strike on the board once a player wins. CSS styling of the width of the strike line also helps.